ARP cache poisoning + DNS spoofing
Implemented ARP cache poisoning. The attacker can obtain a MitM position between a victim and a server. The packets that pass through the attacker are logged in a .pcap file.

Implemented a DNS-spoofing variant. The attacker poisons the ARP of the victim, making the victim think that the attacker is the local DNS. Requests made to specific domains can then be redirected to other IPs by the attacker, while requests to other domains are passed on to the actual local DNS.

Co-Authored-By: rtraitor <rtraitor@users.noreply.github.com>